Reorganize modules to separate api and implementation
