+++ farao-dichotomy-api/checkstyle.xml
