+++ farao-dichotomy-api/src/main/java/com/farao_community/farao/dichotomy/DichotomyEngine.java
