+++ farao-dichotomy-api/src/test/java/com/farao_community/farao/dichotomy/ExampleValidationStrategy.java
