+++ farao-dichotomy-rao-impl/pom.xml
@@ -5,12 +5,12 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.farao-community.farao</groupId>
-    <artifactId>farao-dichotomy</artifactId>
+    <artifactId>farao-dichotomy-rao-impl</artifactId>
     <version>2.1.0-SNAPSHOT</version>
     <packaging>jar</packaging>
 
-    <name>FARAO dichotomy engine</name>
-    <description>A dichotomy engine for process based on remedial actions optimisation and security analysis</description>
+    <name>FARAO dichotomy implementation for RAO</name>
+    <description>Basic implementation set for dichotomy use with FARAO</description>
     <url>http://www.farao-community.com</url>
 
     <licenses>
@@ -36,12 +36,9 @@
     </developers>
 
     <properties>
-        <!-- Project properties -->
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <java.version>11</java.version>
         <maven.compiler.source>11</maven.compiler.source>
+        <maven.compiler.target>11</maven.compiler.target>
 
-        <!-- Dependencies versions -->
         <junit.jupiter.version>RELEASE</junit.jupiter.version>
         <maven.checkstyle.version>3.1.1</maven.checkstyle.version>
         <maven.compiler.version>3.8.1</maven.compiler.version>
@@ -53,10 +50,9 @@
         <maven.surefire.version>2.22.2</maven.surefire.version>
         <slf4j.version>1.7.30</slf4j.version>
 
-        <!-- Business dependencies -->
-        <farao.core.version>3.0.0</farao.core.version>
+        <farao.dichotomy.version>2.1.0-SNAPSHOT</farao.dichotomy.version>
+        <farao.core.version>3.1.0</farao.core.version>
         <powsybl.core.version>4.2.0</powsybl.core.version>
-        <powsybl.rte.core.version>3.2.0</powsybl.rte.core.version>
     </properties>
 
     <build>
@@ -200,11 +196,10 @@
     </profiles>
 
     <dependencies>
-        <!-- Compile dependencies -->
         <dependency>
             <groupId>com.farao-community.farao</groupId>
-            <artifactId>farao-commons</artifactId>
-            <version>${farao.core.version}</version>
+            <artifactId>farao-dichotomy-api</artifactId>
+            <version>${farao.dichotomy.version}</version>
         </dependency>
         <dependency>
             <groupId>com.farao-community.farao</groupId>
@@ -213,7 +208,7 @@
         </dependency>
         <dependency>
             <groupId>com.farao-community.farao</groupId>
-            <artifactId>farao-crac-io-json</artifactId>
+            <artifactId>farao-rao-result-api</artifactId>
             <version>${farao.core.version}</version>
         </dependency>
         <dependency>
@@ -228,18 +223,8 @@
         </dependency>
         <dependency>
             <groupId>com.powsybl</groupId>
-            <artifactId>powsybl-iidm-api</artifactId>
-            <version>${powsybl.core.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.powsybl</groupId>
             <artifactId>powsybl-iidm-xml-converter</artifactId>
             <version>${powsybl.core.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>slf4j-api</artifactId>
-            <version>${slf4j.version}</version>
         </dependency>
 
         <!-- Test dependencies -->
@@ -253,13 +238,6 @@
             <groupId>com.powsybl</groupId>
             <artifactId>powsybl-config-test</artifactId>
             <version>${powsybl.core.version}</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>com.powsybl</groupId>
-            <artifactId>powsybl-iidm-impl</artifactId>
-            <version>${powsybl.core.version}</version>
-            <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>org.junit.jupiter</groupId>
@@ -274,4 +252,5 @@
             <scope>test</scope>
         </dependency>
     </dependencies>
+
 </project>
