+++ farao-dichotomy-rao-impl/src/main/java/com/farao_community/farao/dichotomy/RaoStepResult.java
@@ -4,12 +4,11 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.farao_community.farao.dichotomy.rao;
+package com.farao_community.farao.dichotomy;
 
 import com.farao_community.farao.data.rao_result_api.ComputationStatus;
 import com.farao_community.farao.data.rao_result_api.OptimizationState;
 import com.farao_community.farao.data.rao_result_api.RaoResult;
-import com.farao_community.farao.dichotomy.StepResult;
 
 /**
  * @author Sebastien Murgey {@literal <sebastien.murgey at rte-france.com>}
