+++ farao-dichotomy-rao-impl/src/test/java/com/farao_community/farao/dichotomy/BusinessTestCaseRaoResult.java
