+++ farao-dichotomy-rao-impl/src/test/resources/simplelogger.properties
